OGM-145 Moving AssociationKind to o.h.o.model.key.spi as it's only referred to by metadata
